More messing around with path finding. Grid displays correctly Unit doesnt move correctly Will work on fixing movement over the coming days

+++ TBS_Test/Assets/Pathfinding/Scenes/Scripts/SetGrid.cs
@@ -3,8 +3,8 @@ using System.Collections;
 
 public class SetGrid : MonoBehaviour {
 
-    public int X = 1;
-    public int Y = 1;
+    public float X = 1;
+    public float Y = 1;
 
     void Start()
     {
